refactor: split timeline snippets into separate components (#174)
* extract timeline toolbar to separate component

* use wiki component directly without snippet

* inline thread date

* extract ThreadsTab to component and inline chat snippet
